Fix: Update stack outputs only on successful deployments
If an update fails we should not overwrite the stack's outputs.

Implementation note: when we receive a RegisterResourceOutputsEvent for the stack resource, we defer processing it until the end of the deployment. If the deployment was successful, we update the stack outputs. If there was an error, we keep the old stack outputs.

With this, stack outputs are determined entirely by the last successful run.